Dockerize the project
Use docker to make it easier to set up a local development. Note that
the focus of this dockerization effort has been for local development,
not deployment. There are changes that may need to be made for
deployment, such as adding the php files to the image rather than
mounting a directory.

This also required changes to how the project is configured, especially
around secrets.