Add admin page to manage roles
This allows them to create and delete roles, and set what permissions
those roles have.

There is currently no page to assign roles to members, that has to be
done directly in the database. We still need to decide whether roles
should be managed in the portal, or somewhere outside of it (e.g.
Clyde).

This work required changing the database structure a little to encode
the permissions in a table that could be queried. In the process of
doing this, it became apparent that the default roles duplicated admin
and moderator twice. This had to be cleaned up, and the duplicates
removed.