Add ability to preview roles
People with the manage-roles permission can now add a `role=<name>`
query parameter to preview what a page looks like with the given role.

In order to not completely kill performance, as this adds several extra
db queries to the permissions check, we now resolve all the permissions
up-front and store it in the session object. This is a general
performance win, as previously every permissions check required 1-3
database queries.