Apply some basic branding
Apply some basic Glasgow branding to the portal pages. This hasn't been
run by any of the relevant teams, I just copied and adapted some styles
from the existing website. But it should do for now as something that is
a little closer to what we'll end up with in terms of look and feel for
when we start showing other people.