Remove eastercon resources and settle on initial Glasgow cards
Remove the last of the eastercon specific resources and settle on the
list of cards we're using for Glasgow. This is subject to change, but is
the current list we have.

This switches to defining the cards using an array and templatign them
rather than writing the HTML directly. This makes it a bit easier to see
what the cards are at a glance, and make changes across all cards.

It also adds the notion that some links can be cards for some people,
and just "Other" links for others. An example of this is childcare. The
idea is that anyone who has a child membership in their transaction will
be shown the childcare link in the "Other" section, but people who have
explicitly requested childcare will be shown a full-blown card.