Add script to generate the SQL and invite CSV
This script will be used just before we open up the portal to seed the
database with the current membership.

It also takes advantage of the extra data we have from Clyde to add
roles for WSFS voting rights and childcare.

In future we will need this to generate RCE links and merge in the data
from Planorama for programme participants.