Add "Skip log in" link
Add the option to skip the log in. This will allow people who can't or
don't want to log in (e.g. people under 16) to still have access to the
useful links in the portal.

Places that require authentication, such as the RCE or Discord links,
still require you to log in.