Add API_KEYS to config
Add API_KEYS to config for use by the Discord API. This was just
something that got missed in the initial conversion of how config is
configured in the new dockerized world.

This new config has a slight tweak. Before we would allow multiple keys
to be configured for the same client name to allow rotation. For
simplicity, we have gotten rid of that. You can just use a new client
name for the new key.