Add /api/verify-discord-id to let members link discord ids
Add a new endpoint to allow users to link their discord identity with
their membership.

The process of joining the discord through the portal has two discrete
steps:

1. Get their discord account and associate it with their membership
2. Invite them to the discord

While 95% of the time both steps happen with the same discord account,
certain circumstances mean step 1 and 2 can happen with different
discord accounts. The most common scenario is they don't realise they
have a discord account, so create a new one in step 1, and then in step
2 it loads the discord app logged into their old account.

This means we can have people joining the discord and we don't know
which membership their discord account links to. To fill that gap, we
now send a URL in the /api/check-discord-ids reponse instead of null.
That url can be used to link the given discord account with the logged
in membership. The Discord bot Watson then sends that link to them in a
private thread.